server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyJWT==2.8.0
python-decouple==3.8
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
